linux-kernel: Add sync option to qemu-start Add support to start QEMU with original rootfs and kernel. This is useful when a change in rootfs needs to be sync-ed across QEMU instances.

+++ linux-kernel/misc/qemu-start.sh
@@ -17,42 +17,54 @@ QEMU_UTIL_DIR=/home/$USER/Workspace/qemu-utils
 QEMU_MAX_INSTANCE=4
 QEMU_INSTANCE=0
 
-# assign an instance number for the QEMU instance
-for (( i=1; i<=$QEMU_MAX_INSTANCE; i++ ))
-do
-	if [ -d "$QEMU_SIM_DIR/$i" ]
-	then
-		echo "Already running QEMU instance $i: $QEMU_SIM_DIR/$i"
-	else
-		echo "Creating QEMU instance $i: $QEMU_SIM_DIR/$i"
-		QEMU_INSTANCE=$i
-		break
-	fi
-done
-
-# create a QEMU instance if the max. number of possible instances is not reached yet
-if [[ $QEMU_INSTANCE -eq 0 ]]
+if [[ $1 = "sync" ]]
 then
-	echo "Already running max. number[$QEMU_MAX_INSTANCE] of allowed QEMU instances!"
-	echo "If you are not running QEMU and still see this message,"
-	echo "then there may be leftover files from a previous run which did not exit gracefully."
-	echo "You can clean these files up at $QEMU_SIM_DIR to resolve this issue:"
-	echo " "
-	echo "rm -rf $QEMU_SIM_DIR/*"
-	echo " "
-else
-	mkdir -p $QEMU_SIM_DIR/$QEMU_INSTANCE
-	cp $QEMU_UTIL_DIR/bzImage $QEMU_SIM_DIR/$QEMU_INSTANCE/
-	cp $QEMU_UTIL_DIR/disk.img $QEMU_SIM_DIR/$QEMU_INSTANCE/
-	cp $QEMU_UTIL_DIR/qemu-ifup.sh $QEMU_SIM_DIR/$QEMU_INSTANCE/
-
+	# run a QEMU instance using the original rootfs and kernel
+	# useful to sync. changes in rootfs
 	sudo qemu-system-x86_64 -enable-kvm -nographic \
 		-cpu host -machine q35 -smp 4 -m 512M \
-		-kernel $QEMU_SIM_DIR/$QEMU_INSTANCE/bzImage \
-		-hda $QEMU_SIM_DIR/$QEMU_INSTANCE/disk.img \
+		-kernel $QEMU_UTIL_DIR/bzImage \
+		-hda $QEMU_UTIL_DIR/disk.img \
 		-append "root=/dev/sda rw console=ttyS0" \
-		-nic tap,model=e1000,script=$QEMU_SIM_DIR/$QEMU_INSTANCE/qemu-ifup.sh
+		-nic tap,model=e1000,script=$QEMU_UTIL_DIR/qemu-ifup.sh
+else
+	# assign an instance number for the QEMU instance
+	for (( i=1; i<=$QEMU_MAX_INSTANCE; i++ ))
+	do
+		if [ -d "$QEMU_SIM_DIR/$i" ]
+		then
+			echo "Already running QEMU instance $i: $QEMU_SIM_DIR/$i"
+		else
+			echo "Creating QEMU instance $i: $QEMU_SIM_DIR/$i"
+			QEMU_INSTANCE=$i
+			break
+		fi
+	done
 
-	echo "Cleaning up $QEMU_SIM_DIR/$QEMU_INSTANCE"
-	rm -rf $QEMU_SIM_DIR/$QEMU_INSTANCE
+	# create a QEMU instance if the max. number of possible instances is not reached yet
+	if [[ $QEMU_INSTANCE -eq 0 ]]
+	then
+		echo "Already running max. number[$QEMU_MAX_INSTANCE] of allowed QEMU instances!"
+		echo "If you are not running QEMU and still see this message,"
+		echo "then there may be leftover files from a previous run which did not exit gracefully."
+		echo "You can clean these files up at $QEMU_SIM_DIR to resolve this issue:"
+		echo " "
+		echo "rm -rf $QEMU_SIM_DIR/*"
+		echo " "
+	else
+		mkdir -p $QEMU_SIM_DIR/$QEMU_INSTANCE
+		cp $QEMU_UTIL_DIR/bzImage $QEMU_SIM_DIR/$QEMU_INSTANCE/
+		cp $QEMU_UTIL_DIR/disk.img $QEMU_SIM_DIR/$QEMU_INSTANCE/
+		cp $QEMU_UTIL_DIR/qemu-ifup.sh $QEMU_SIM_DIR/$QEMU_INSTANCE/
+
+		sudo qemu-system-x86_64 -enable-kvm -nographic \
+			-cpu host -machine q35 -smp 4 -m 512M \
+			-kernel $QEMU_SIM_DIR/$QEMU_INSTANCE/bzImage \
+			-hda $QEMU_SIM_DIR/$QEMU_INSTANCE/disk.img \
+			-append "root=/dev/sda rw console=ttyS0" \
+			-nic tap,model=e1000,script=$QEMU_SIM_DIR/$QEMU_INSTANCE/qemu-ifup.sh
+
+		echo "Cleaning up $QEMU_SIM_DIR/$QEMU_INSTANCE"
+		rm -rf $QEMU_SIM_DIR/$QEMU_INSTANCE
+	fi
 fi
